#96: 코드 분리
