--- ToDo-Garden/ToDoGardenUI/Sources/ToDoGardenUIComponent/RowStyle.swift
@@ -1,424 +1,0 @@
-import Combine
-import ToDoGardenUIResource
-import UIKit
-
-extension Styled {
-  final public class Row: UIView {
-    public var iconImage: UIImage? {
-      get {
-        if let image = self.configutration.profileModel?.image {
-          return image
-        }
-        return nil
-      }
-      set {
-        if var model = self.configutration.profileModel, let newValue {
-          model.image = newValue
-          self.configutration = .profile(model)
-        }
-      }
-    }
-    
-    @Published var configutration: Configuration
-    private var cancellables: Set<AnyCancellable> = []
-    
-    public init(configuration: Configuration) {
-      self.configutration = configuration
-      super.init(frame: CGRect.zero)
-      self.build()
-    }
-    
-    @available(*, unavailable)
-    public required init?(coder: NSCoder) {
-      fatalError("init(coder:) has not been implemented")
-    }
-    
-    deinit {
-      self.cancellables.removeAll()
-    }
-    
-    private func build() {
-      let stack = UIStackView(frame: CGRect.zero)
-      switch self.configutration {
-      case let Configuration.profile(profileModel):
-        self.buildProfileStyle(stack: stack, model: profileModel)
-      case let Configuration.listPrimary(listPrimaryModel):
-        self.buildListPrimaryStyle(stack: stack, model: listPrimaryModel)
-      case let Configuration.todoList(todoListModel):
-        self.buildTodoListStyle(stack: stack, model: todoListModel)
-      }
-    }
-  }
-}
-
-extension Styled.Row {
-  public enum Configuration {
-    var profileModel: ProfileModel? {
-      if case let Self.profile(model) = self {
-        return model
-      }
-      return nil
-    }
-    
-    var listPrimaryModel: ListPrimaryModel? {
-      if case let Self.listPrimary(model) = self {
-        return model
-      }
-      return nil
-    }
-    var todoListModel: TodoListModel? {
-      if case let Self.todoList(model) = self {
-        return model
-      }
-      return nil
-    }
-    
-    case profile(ProfileModel)
-    case listPrimary(ListPrimaryModel)
-    case todoList(TodoListModel)
-  }
-}
-
-extension Styled.Row.Configuration {
-  public struct ProfileModel: Equatable {
-    public static func primary(
-      image: UIImage = UIImage.defaultProfileImage,
-      title: String,
-      description: String
-    ) -> Self {
-      return Self(
-        image: image,
-        title: title,
-        titleFont: UIFont.pretendardHeadBold,
-        description: description,
-        descriptionFont: UIFont.pretendardDetailLight,
-        axis: .vertical
-      )
-    }
-    
-    public static func gardenInfo(
-      image: UIImage = UIImage.defaultProfileImage,
-      title: String,
-      description: String
-    ) -> Self {
-      return Self(
-        image: image,
-        title: title,
-        titleFont: UIFont.pretendardBodySemiBold,
-        description: description,
-        descriptionFont: UIFont.pretendardBodyMedium,
-        axis: .horizontal
-      )
-    }
-    var image: UIImage = UIImage.defaultFriendProfileImage
-    var title: String
-    var titleFont: UIFont
-    var description: String
-    var descriptionFont: UIFont
-    var axis: NSLayoutConstraint.Axis
-  }
-  
-  public struct ListPrimaryModel: Equatable {
-    let title: String
-    let color: UIColor
-  }
-  
-  public struct TodoListModel: Equatable {
-    public static let empty = Self()
-    public var text: String?
-    public var isSelected: Bool
-    
-    public init(text: String? = nil, isSelected: Bool = false) {
-      self.text = text
-      self.isSelected = isSelected
-    }
-  }
-}
-
-// MARK: - Shared View Logic
-extension Styled.Row {
-  private func buildStack(
-    stack: UIStackView,
-    axis: NSLayoutConstraint.Axis = NSLayoutConstraint.Axis.horizontal,
-    edgeInsets: NSDirectionalEdgeInsets
-  ) {
-    stack.alignment = .center
-    stack.axis = axis
-    stack.isLayoutMarginsRelativeArrangement = true
-    stack.directionalLayoutMargins = edgeInsets
-    self.addSubview(stack)
-    stack.usingAutolayout()
-    NSLayoutConstraint.activate([
-      stack.topAnchor.constraint(equalTo: self.topAnchor),
-      stack.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-      stack.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-      stack.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-    ])
-  }
-  
-  @discardableResult
-  private func buildImageView(stack: UIStackView, image: UIImage, size: CGSize) -> UIImageView {
-    let imageView = UIImageView(image: image)
-    imageView.usingAutolayout()
-    NSLayoutConstraint.activate([
-      imageView.widthAnchor.constraint(equalToConstant: size.width),
-      imageView.heightAnchor.constraint(equalToConstant: size.height)
-    ])
-    stack.addArrangedSubview(imageView)
-    return imageView
-  }
-  
-  @discardableResult
-  private func buildTextLabel(stack: UIStackView, text: String, font: UIFont, textColor: UIColor) -> UILabel {
-    let label = UILabel()
-    label.text = text
-    label.font = font
-    label.textColor = textColor
-    stack.addArrangedSubview(label)
-    return label
-  }
-}
-
-// MARK: - ProfileStyle
-extension Styled.Row {
-  private func buildProfileStyle(stack: UIStackView, model: Configuration.ProfileModel) {
-    self.buildProfileStyleStack(stack: stack)
-    let imageView = self.buildImageView(
-      stack: stack,
-      image: model.image,
-      size: CGSize(width: 55, height: 55)
-    )
-    self.buildInnerStack(stack: stack, model: model)
-    if model.axis == NSLayoutConstraint.Axis.vertical {
-      stack.addSpacing()
-    }
-    self.buildImageView(
-      stack: stack,
-      image: UIImage.forwardButtonImage,
-      size: CGSize(width: 24, height: 24)
-    )
-    self.bindingProfileImageState(imageView: imageView)
-  }
-  
-  private func buildProfileStyleStack(stack: UIStackView) {
-    stack.isLayoutMarginsRelativeArrangement = true
-    stack.spacing = 15
-    self.buildStack(
-      stack: stack,
-      edgeInsets: NSDirectionalEdgeInsets(
-        top: 8,
-        leading: 16,
-        bottom: 8,
-        trailing: 36
-      )
-    )
-  }
-  
-  private func buildInnerStack(stack: UIStackView, model: Configuration.ProfileModel) {
-    let innerStack = UIStackView(frame: CGRect.zero)
-    innerStack.axis = model.axis
-    let titleLabel = self.buildTextLabel(
-      stack: innerStack,
-      text: model.title,
-      font: model.titleFont,
-      textColor: .toDoGardenGreenDark
-    )
-    addConditionalSpacing(innerStack, axis: model.axis)
-    let descriptionLabel = self.buildTextLabel(
-      stack: innerStack,
-      text: model.description,
-      font: model.descriptionFont,
-      textColor: .toDoGardenGreenDark
-    )
-    stack.addArrangedSubview(innerStack)
-    bindingProfileInnerTitleState(
-      titleLabel: titleLabel,
-      descriptionLabel: descriptionLabel
-    )
-  }
-  
-  private func addConditionalSpacing(_ stack: UIStackView, axis: NSLayoutConstraint.Axis) {
-    switch axis {
-    case NSLayoutConstraint.Axis.vertical:
-      stack.spacing = 3
-    case NSLayoutConstraint.Axis.horizontal:
-      stack.addSpacing()
-    @unknown default:
-      break
-    }
-  }
-  
-  private func bindingProfileImageState(imageView: UIImageView) {
-    self.$configutration
-      .map(\.profileModel?.image)
-      .removeDuplicates()
-      .sink { [weak imageView] image in
-        imageView?.image = image
-      }
-      .store(in: &cancellables)
-  }
-  
-  private func bindingProfileInnerTitleState(
-    titleLabel: UILabel,
-    descriptionLabel: UILabel
-  ) {
-    self.$configutration
-      .map(\.profileModel?.description)
-      .removeDuplicates()
-      .sink { [weak descriptionLabel] text in
-        descriptionLabel?.text = text
-      }
-      .store(in: &cancellables)
-    self.$configutration
-      .map(\.profileModel?.title)
-      .removeDuplicates()
-      .sink { [weak titleLabel] text in
-        titleLabel?.text = text
-      }
-      .store(in: &cancellables)
-  }
-}
-
-// MARK: - ListPrimary
-extension Styled.Row {
-  private func buildListPrimaryStyle(stack: UIStackView, model: Configuration.ListPrimaryModel) {
-    stack.alignment = UIStackView.Alignment.center
-    self.buildStack(
-      stack: stack,
-      edgeInsets: NSDirectionalEdgeInsets(
-        top: 0,
-        leading: 14,
-        bottom: 0,
-        trailing: 14
-      )
-    )
-    // MARK: - TODO: 레이블 집어넣기
-    
-    stack.addSpacing()
-    self.buildColorView(stack: stack, color: model.color)
-  }
-  
-  private func buildColorView(stack: UIStackView, color: UIColor) {
-    let colorView = UIView()
-    colorView.backgroundColor = color
-    colorView.layer.cornerRadius = 12
-    colorView.usingAutolayout()
-    NSLayoutConstraint.activate([
-      colorView.widthAnchor.constraint(equalToConstant: 24),
-      colorView.heightAnchor.constraint(equalToConstant: 24)
-    ])
-    stack.addArrangedSubview(colorView)
-  }
-}
-
-extension Styled.Row {
-  private func buildTodoListStyle(stack: UIStackView, model: Configuration.TodoListModel) {
-    self.buildStack(
-      stack: stack,
-      edgeInsets: NSDirectionalEdgeInsets(top: 12, leading: 41, bottom: 12, trailing: 0)
-    )
-    let textField = self.buildTextField(stack: stack, text: model.text)
-    self.buildButton(stack: stack, textField: textField)
-    stack.addSpacing(8)
-    stack.addArrangedSubview(textField)
-    stack.addSpacing()
-  }
-  
-  private func buildButton(stack: UIStackView, textField: UITextField) {
-    let button = UIButton(
-      configuration: .plain(),
-      primaryAction: UIAction { [weak self, weak textField] action in
-        guard
-          let button = action.sender as? UIButton
-        else { return }
-        self?.updateTextField(textField, buttonSelected: button.isSelected)
-        self?.configutration.todoListModel.map { model in
-          var copy = model
-          copy.isSelected = button.isSelected
-          self?.configutration = .todoList(copy)
-        }
-      }
-    )
-    buildButtonConfiguration(button)
-    builButtonLayout(button, stack: stack)
-  }
-  
-  private func buildButtonConfiguration(_ button: UIButton) {
-    button.configuration?.baseForegroundColor = UIColor.toDoGardenRed
-    button.configurationUpdateHandler = { button in
-      switch button.state {
-      case .selected, .highlighted:
-        let image = UIImage(systemName: "checkmark.square.fill")
-        button.configuration?.image = image
-      default:
-        button.configuration?.image = UIImage(systemName: "square")
-      }
-    }
-    button.changesSelectionAsPrimaryAction = true
-  }
-  
-  private func builButtonLayout(_ button: UIButton, stack: UIStackView) {
-    button.usingAutolayout()
-    NSLayoutConstraint.activate([
-      button.widthAnchor.constraint(equalToConstant: 18),
-      button.heightAnchor.constraint(equalToConstant: 18)
-    ])
-    stack.addArrangedSubview(button)
-  }
-  
-  private func buildTextField(stack: UIStackView, text: String?) -> UITextField {
-    let textField = Styled.UITextField(configuration: .groupEdit(.todoList))
-    textField.text = text
-    let action = UIAction { [weak self] action in
-      if let textField = action.sender as? UITextField {
-        self?.configutration.todoListModel
-          .map { model in
-            var copy = model
-            copy.text = textField.text
-            self?.configutration = .todoList(copy)
-          }
-      }
-    }
-    textField.addAction(action, for: .editingChanged)
-    textField.usingAutolayout()
-    NSLayoutConstraint.activate([
-      textField.widthAnchor.constraint(equalToConstant: 200)
-    ])
-    
-    return textField
-  }
-  
-  private func updateTextField(_ textField: UITextField?, buttonSelected: Bool) {
-    textField?.isEnabled = !buttonSelected
-    textField?.textColor = buttonSelected ? UIColor.gray : UIColor.black
-    if buttonSelected {
-      let attributeString = NSMutableAttributedString(string: textField?.text ?? "")
-      attributeString
-        .addAttribute(
-          NSAttributedString.Key.strikethroughStyle,
-          value: 1,
-          range: NSRange(location: 0, length: attributeString.length)
-        )
-      textField?.attributedText = attributeString
-    } else {
-      let temp = textField?.text
-      textField?.attributedText = nil
-      textField?.text = temp
-    }
-  }
-}
-
-#if DEBUG
-@available(iOS 17.0, *)
-#Preview {
-  let stack = UIStackView()
-  stack.axis = .vertical
-  
-  let view = Styled.Row(configuration: .profile(.primary(title: "배고픔", description: "울트라 배고픔")))
-  stack.addArrangedSubview(view)
-  
-  let view2 = Styled.Row(configuration: .profile(.gardenInfo(title: "윤강운", description: "강운이는 크다.")))
-  stack.addArrangedSubview(view2)
-  return stack
-}
-#endif
